user

+++ src/main/java/gestion_user/service/UserService.java
@@ -1,8 +1,9 @@
 package gestion_user.service;
 
-import gestion_user.entities.Adherent;
+
 import gestion_user.entities.Role;
 import gestion_user.entities.Sexe;
+import gestion_user.entities.User;
 import utils.DataSource;
 
 import java.sql.*;
@@ -10,19 +11,19 @@ import java.util.ArrayList;
 import java.util.List;
 
 
-public class AdherentService implements IService<Adherent> {
+public class UserService implements IService<User> {
 
     private Connection conn;
     private Statement ste;
     private PreparedStatement pst;
 
-    public AdherentService() {
+    public UserService() {
         conn = DataSource.getInstance().getCnx();
     }
 
     @Override
-    public void add(Adherent c) {
-        String requete = "INSERT INTO user (nom, prenom, mot_de_passe, email, date_de_naissance, sexe, role) VALUES (?, ?, ?, ?, ?, ?, ?)";
+    public void addAdherent(User c) {
+        String requete = "INSERT INTO user (nom, prenom, mot_de_passe, email, date_de_naissance, sexe, role,salaire, id_bilan_financier) VALUES (?, ?, ?, ?, ?, ?, ?,?,?)";
         try {
             pst = conn.prepareStatement(requete);
             pst.setString(1, c.getNom());
@@ -31,7 +32,9 @@ public class AdherentService implements IService<Adherent> {
             pst.setString(4, c.getEmail());
             pst.setDate(5, c.getDate_de_naissance());
             pst.setString(6, c.getSexe().toString());
-            pst.setString(7, Role.Adherent.toString());
+            pst.setString(7, c.getRole().toString());
+            pst.setDouble(8,c.getSalaire());
+            pst.setInt(9,c.getId_bilan_financier());
 
             pst.executeUpdate();
 
@@ -44,8 +47,8 @@ public class AdherentService implements IService<Adherent> {
 
 
     @Override
-    public void update(int id,String role, Adherent c) {
-        String requete = "UPDATE user SET nom = ?, prenom = ?, mot_de_passe = ?, email = ?, date_de_naissance = ?, sexe = ?  WHERE id = ? AND role= ?";
+    public void updateAdherent(int id, User c) {
+        String requete = "UPDATE user SET nom = ?, prenom = ?, mot_de_passe = ?, email = ?, date_de_naissance = ?, sexe = ?, role= ?, salaire = ?,id_bilan_financier = ?  WHERE id = ?";
 
         try (PreparedStatement pst = conn.prepareStatement(requete)) {
 
@@ -55,13 +58,15 @@ public class AdherentService implements IService<Adherent> {
             pst.setString(4, c.getEmail());
             pst.setDate(5, c.getDate_de_naissance());
             pst.setString(6, c.getSexe().toString());
-            pst.setInt(7,id);
-            pst.setString(8,role);
+            pst.setString(7,c.getRole().toString());
+            pst.setDouble(8, c.getSalaire());
+            pst.setInt(9,c.getId_bilan_financier());
+            pst.setInt(10,id);
             int test= pst.executeUpdate();
             if (test == 0) {
                 System.out.println("Colone n'est pas mettre à jour " + id);
             } else {
-                System.out.println("Adherent trouvé " + id + " est updaté.");
+                System.out.println("Colone trouvé " + id + " est updaté.");
             }
 
         } catch(SQLException e){
@@ -70,9 +75,9 @@ public class AdherentService implements IService<Adherent> {
     }
 
     @Override
-    public  List<Adherent> readAll() {
-        String requete = "select * from user where role='Adherent'";
-        List<Adherent> list = new ArrayList<>();
+    public  List<User> readAllAdherent() {
+        String requete = "select * from user";
+        List<User> list = new ArrayList<>();
         try {
             ste = conn.createStatement();
             ResultSet rs = ste.executeQuery(requete);
@@ -80,8 +85,12 @@ public class AdherentService implements IService<Adherent> {
                 Sexe sexe = Sexe.valueOf(rs.getString(7));
                 Role role = Role.valueOf(rs.getString(8));
 
-                Adherent ad = new Adherent( rs.getString(2), rs.getString(3), rs.getString(4), rs.getString(5), rs.getDate(6), sexe,role);
-                    ad.setId(rs.getInt(1));
+                User ad = new User( rs.getString(2), rs.getString(3), rs.getString(4), rs.getString(5), rs.getDate(6), sexe,role);
+                ad.setId(rs.getInt(1));
+                ad.setSalaire(rs.getDouble(9));
+                ad.setId_bilan_financier(rs.getInt(10));
+
+
                     list.add(ad);
                                 }
         } catch (SQLException e) {
@@ -91,8 +100,8 @@ public class AdherentService implements IService<Adherent> {
     }
 
     @Override
-    public void delete(int id) {
-        String requete = "DELETE FROM user WHERE id = ?";
+    public void deleteAdherent(int id) {
+        String requete = "DELETE FROM user WHERE id = ? and role='Adherent'";
         try{
             pst = conn.prepareStatement(requete);
             pst.setInt(1 , id);
@@ -100,7 +109,7 @@ public class AdherentService implements IService<Adherent> {
             if (test == 0) {
                 System.out.println("N'y a aucun Adherent avec ce id: " + id);
             } else {
-                System.out.println("Adhrent trouvé " + id + " a été supprimé.");
+                System.out.println("Adherent trouvé " + id + " a été supprimé.");
             }
         } catch(SQLException e){
             throw new RuntimeException(e);
@@ -108,17 +117,17 @@ public class AdherentService implements IService<Adherent> {
 
 
 
-    @Override
-    public Adherent readById(int id) {
-        String requete = "SELECT * FROM user WHERE id = ?";
-        Adherent adherent = null;
+   /* @Override
+    public Adherent readAdherentById(int id) {
+        String requete = "SELECT * FROM user WHERE id = ? and role='Adherent'";
+        User user = null;
         try (PreparedStatement pst = conn.prepareStatement(requete)) {
             pst.setInt(1, id);
             try (ResultSet rs = pst.executeQuery()) {
                 if (rs.next()) {
                     Sexe sexe = Sexe.valueOf(rs.getString(7));
                     Role role = Role.valueOf(rs.getString(8));
-                    adherent = new Adherent(
+                    user = new User(
                             rs.getInt(1),
                             rs.getString(2),
                             rs.getString(3),
@@ -136,7 +145,7 @@ public class AdherentService implements IService<Adherent> {
             throw new RuntimeException(e);
 
         }
-        return adherent;
-    }
+        return user;
+    }*/
 
 }
